refactor: move Worker Request Handler of module
