+++ jina/serve/runtimes/worker/__init__.py
@@ -13,7 +13,7 @@ from jina.proto import jina_pb2, jina_pb2_grpc
 from jina.serve.instrumentation import MetricsTimer
 from jina.serve.runtimes.asyncio import AsyncNewLoopRuntime
 from jina.serve.runtimes.helper import _get_grpc_server_options
-from jina.serve.runtimes.request_handlers.worker_request_handler import (
+from jina.serve.runtimes.worker.worker_request_handler import (
     WorkerRequestHandler,
 )
 from jina.types.request.data import DataRequest
