+++ tests/unit/serve/runtimes/request_handlers/test_worker_request_handler.py
@@ -5,7 +5,7 @@ from jina import Executor, requests
 from jina.clients.request import request_generator
 from jina.logging.logger import JinaLogger
 from jina.parsers import set_pod_parser
-from jina.serve.runtimes.request_handlers.worker_request_handler import WorkerRequestHandler
+from jina.serve.runtimes.worker.worker_request_handler import WorkerRequestHandler
 
 
 class NewDocsExecutor(Executor):
