+++ tests/unit/serve/runtimes/worker/test_worker_runtime.py
@@ -1,6 +1,5 @@
 import asyncio
 import multiprocessing
-import os
 import socket
 import time
 from multiprocessing import Process
@@ -17,7 +16,7 @@ from jina.parsers import set_pod_parser
 from jina.proto import jina_pb2, jina_pb2_grpc
 from jina.serve.networking import GrpcConnectionPool
 from jina.serve.runtimes.asyncio import AsyncNewLoopRuntime
-from jina.serve.runtimes.request_handlers.worker_request_handler import WorkerRequestHandler
+from jina.serve.runtimes.worker.worker_request_handler import WorkerRequestHandler
 from jina.serve.runtimes.worker import WorkerRuntime
